requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.13.0
scikit-learn>=1.3.0
scipy>=1.10.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
einops>=0.7.0
jupyter>=1.0.0
ipywidgets>=8.0.0